requirements.txt
pyobjc==10.2
pyobjc-framework-Cocoa==10.2
py2app>=0.28.8
Pillow>=10.0.0
setuptools<81
